+++ tests/test_telephony_route_list.py
@@ -204,7 +204,7 @@ class TestNumbers(TestCaseWithLog):
         """
         # get phone numbers in target location
         available_tns = [n.phone_number for n in self.numbers
-                         if n.location.location_id == self.target_location_info.location.location_id
+                         if n.location.id == self.target_location_info.location.location_id
                          and not n.main_number and n.owner is None]
         if len(available_tns) < tn_count:
             with self.no_log():